requirements.txt
openai
python-dotenv
chromadb
uuid
datasets
rank_bm25
tavily-python
nltk
google-generativeai
qdrant-client